backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-docx==1.1.0
docxtpl==0.16.7
pyjwt==2.8.0
email-validator==2.1.0
httpx==0.27.0
alembic==1.13.1
num2words==0.5.13
